refactor harvest related methods into module
Moved the harvest related methods out of lib/new_relic/agent/agent.rb
into its own module and file
